fix(quality): address static analysis comments
