+++ lemma-backend/app/modules/agent/services/agent_runner_service.py
@@ -181,7 +181,8 @@ class AgentRunObserver(Protocol):
         self,
         conversation: Conversation,
         error: Exception,
-    ) -> None: ...
+    ) -> None:
+        raise NotImplementedError
 
 
 class AgentRunnerService:
